Change active site align endpoint to accept custom motifs
