+++ README.md
@@ -37,7 +37,6 @@ This briefly summarizes all API endpoints.
 | HTTP Method | Endpoint | Function |
 |:------------|:---------|:---------|
 | POST | [/align/activesite](#post-align-activesite) | Aligns protein active sites with motifs |
-| POST | [/align/backbone](#post-align-backbone) | NOT IMPLEMENTED. Aligns protein backbones with motifs |
 
 #### Motifs
 
@@ -77,14 +76,6 @@ Example JSON request
 	"filters": []
 }
 ```
-
-
-<a name="post-align-backbone"></a>
-##### POST /align/backbone
-###### Aligns protein backbones with motifs
-
-NOT IMPLEMENTED
-
