+++ src/main/java/org/moltimate/moltimatebackend/controller/TaskController.java
@@ -1,7 +1,7 @@
 package org.moltimate.moltimatebackend.controller;
 
 import lombok.extern.slf4j.Slf4j;
-import org.moltimate.moltimatebackend.service.GenerateMotifService;
+import org.moltimate.moltimatebackend.service.MotifService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -16,13 +16,13 @@ import org.springframework.web.bind.annotation.RestController;
 public class TaskController {
 
     @Autowired
-    private GenerateMotifService generateMotifService;
+    private MotifService motifService;
 
     /**
      * Updates the motif database using the CSA and RCSB PDB
      */
     @RequestMapping(value = "/updatemotifs", method = RequestMethod.GET)
     public void updateMotifs() {
-        generateMotifService.updateMotifs();
+        motifService.updateMotifs();
     }
 }
