+++ src/main/java/org/moltimate/moltimatebackend/response/ActiveSiteAlignmentResponse.java
@@ -14,6 +14,6 @@ import java.util.Map;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class AlignmentResponse {
+public class ActiveSiteAlignmentResponse {
     Map<String, List<Alignment>> alignments;
 }
