+++ src/main/java/org/moltimate/moltimatebackend/service/AlignmentService.java
@@ -7,8 +7,7 @@ import org.moltimate.moltimatebackend.model.Alignment;
 import org.moltimate.moltimatebackend.model.Motif;
 import org.moltimate.moltimatebackend.model.Residue;
 import org.moltimate.moltimatebackend.request.ActiveSiteAlignmentRequest;
-import org.moltimate.moltimatebackend.request.BackboneAlignmentRequest;
-import org.moltimate.moltimatebackend.response.AlignmentResponse;
+import org.moltimate.moltimatebackend.response.ActiveSiteAlignmentResponse;
 import org.moltimate.moltimatebackend.util.AlignmentUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -26,7 +25,7 @@ import java.util.Set;
 import java.util.stream.Collectors;
 
 /**
- * AlignmentService provides a way to align the active sites and backbones of proteins
+ * AlignmentService provides a way to align the active sites of proteins
  */
 @Service
 @Slf4j
@@ -42,16 +41,16 @@ public class AlignmentService {
      * Executes the ActiveSiteAlignmentRequest on protein active sites.
      *
      * @param alignmentRequest The alignment request JSON mapped to an object
-     * @return AlignmentResponse which contains all alignments and their relevant data
+     * @return ActiveSiteAlignmentResponse which contains all alignments and their relevant data
      */
-    public AlignmentResponse alignActiveSites(ActiveSiteAlignmentRequest alignmentRequest) {
+    public ActiveSiteAlignmentResponse alignActiveSites(ActiveSiteAlignmentRequest alignmentRequest) {
         return alignActiveSites(
                 proteinService.queryPdb(alignmentRequest.getPdbIds()),
                 alignmentRequest.getEcNumber()
         );
     }
 
-    private AlignmentResponse alignActiveSites(List<Structure> sourceStructures, String motifEcNumberFilter) {
+    private ActiveSiteAlignmentResponse alignActiveSites(List<Structure> sourceStructures, String motifEcNumberFilter) {
         HashMap<String, List<Alignment>> results = new HashMap<>();
         sourceStructures.forEach(structure -> results.put(structure.getPDBCode(), new ArrayList<>()));
 
@@ -82,8 +81,7 @@ public class AlignmentService {
         }
 
         log.info("Found " + resultsCount + " results");
-
-        return new AlignmentResponse(results);
+        return new ActiveSiteAlignmentResponse(results);
     }
 
     private Alignment alignActiveSites(Structure structure, Motif motif) {
@@ -93,7 +91,6 @@ public class AlignmentService {
         List<Group> seq2 = new ArrayList<>();
 
         Map<Residue, Group> alignmentMapping = new HashMap<>();
-
         Set<Group> found = new HashSet<>();
 
         motif.getActiveSiteResidues()
@@ -159,37 +156,12 @@ public class AlignmentService {
         return null;
     }
 
-    public boolean acceptableDistance(int activeSiteSize, int distance) {
+    private boolean acceptableDistance(int activeSiteSize, int distance) {
         if (activeSiteSize == 2) {
             return false;
         } else if (activeSiteSize >= 3) {
             return distance == 0;
         }
         return distance <= 1;
-    }
-
-    /**
-     * Executes the BackboneAlignmentRequest on protein backbones.
-     *
-     * @param alignmentRequest The alignment request JSON mapped to an object
-     * @return AlignmentResponse which contains all alignments and their relevant data
-     */
-    public AlignmentResponse alignBackbones(BackboneAlignmentRequest alignmentRequest) {
-        return alignBackbones(
-                proteinService.queryPdb(alignmentRequest.getSourcePdbIds()),
-                proteinService.queryPdb(alignmentRequest.getCompareToPdbIds())
-        );
-    }
-
-    private AlignmentResponse alignBackbones(List<Structure> sourceStructures, List<Structure> compareToStructures) {
-        // ... for each sourceStructures, call alignBackbones() against all compareToStructures
-
-        return new AlignmentResponse(/* populate me with a list of Alignments */);
-    }
-
-    private Alignment alignBackbones(Structure structure1, Structure structure2) {
-        // ... logic to perform one backbone alignment between two structures
-
-        return new Alignment();
     }
 }
